requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
beautifulsoup4==4.12.2
tldextract==3.4.4
xgboost==1.7.6
sqlalchemy==2.0.20
python-whois==0.8.0
dnspython==2.6.1
python-Levenshtein==0.23.0
pysafebrowsing==0.1.1
cryptography==41.0.3
urllib3==2.0.4
idna==3.4
tqdm==4.66.1
pyphish==0.1.2
huggingface-hub==0.19.4
joblib==1.3.2
cloudpickle==3.0.0
